Autorun main.py file burned to OpenMV
OpenMV development board is also a usb device, put this main file, when the OpenMV device is connected to the computer through the usb interface, the device will run its own main script. We use a controller to listen to the script to run the device manually, and detach from the plugin provided by OpenMV as well as the IDE, use the command line to run the device.